Move Snapshot and Restore alongside Stash and Unstash
